Moved UserLoginComponent to its proper folder
