+++ src/app/user/user-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AuthGuardGuard } from '../guards/auth-guard.guard';
-import { UserLoginComponent } from '../user-login/user-login.component';
+import { UserLoginComponent } from './components/user-login/user-login.component';
 import { UserRegisterComponent } from './components/user-register/user-register.component';
 
 const routes: Routes = [
